Added a Delete function
Made it so that you can delete the events as well as edit them. This button is right next to the edit button. When the user clicks the delete button, it should pop up and ask them if they are sure they want to delete it. Used javascript validation to do this.